requirements.txt
requests==2.31.0
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
slack_bolt==1.18.1
google-generativeai==0.5.2
google-cloud-aiplatform>=1.38
